chore: move test under correct file
